✨ emacs-everywhere-linux-copy-function emacs-everywhere-paste-cmd
defcustom for alternative methods of handling the clipboard on linux
also added documentation of the added as well as other customizations

Overlaps with issue #28 and maybe also a fix for issue #10

The modifications have at least a few weak spots.
  1. emacs-everywhere-paste-cmd should maybe be a function
  2. organization of added functions and vars
  3. (eval `(... ,@args)) is (apply ...) elsewhere

1. There's a solid argument that it's easier to use as a list:
the intent is clearer and there is little ceremony.
but it is weaker (can't drop into elisp) and inconsistent with
emacs-everywhere-filename-function and emacs-everywhere-linux-copy-function

2. I get the sense things were well organized before I started poking around.
But I'm not deep enough into to understand what should go where.

3. `emacs-everywhere-call` uses
    (apply #'call-process command nil t nil (remq nil args))
  Just saw that. That looks cleaner than using eval. I'm here to
  practice lisp. Got carried away in emacs-everywhere-paste-cmd?